Reorganized [READ DESCRIPTION}
I placed all JavaScript and CSS files within their respective folders in the 'public' directory.
The reason we place JS and CSS files in public is so that the server can serve them statically.
All HTML files will be placed in their respective folders within the project file. So teacher.html would go into the 'teacher' folder and etc.

==LINKING JS OR CSS TO HTML==
To link a JS and CSS file to an HTML file, in the HTML file you would set the href to
href="/css/filename.css" for a CSS file
or
href="/js/filename.js" for a JS file.