Backend for Teacher Page (Not fully done)
The teacher page loads all classes that are tied to the logged in teacher's id. You can then access and submit attendance as needed. That should all work.

However, I wasn't able to complete create and delete class just yet.

You can test this, but first you need the same database. I included Daycare SQL.sql in the project folder which contains all of the sql code to create the database.
Once you run it and create the database, go to localhost:3000/test. This will "log you in" (this basically logs you in to an account I have prepared), you can check if you're logged in if you go to localhost:3000/auth-status.
Afterwards, go to localhost:3000/teacher and you should be able to test it.